# Lint some code plus improve performance with base R and vctrs + reduce pipe