--- a/R/dt_stub_df.R
+++ b/R/dt_stub_df.R
@@ -204,7 +204,7 @@ dt_stub_components <- function(data) {
     stub_components <- c(stub_components, "group_id")
   }
 
-  if (!all(is.na(stub_df[["row_id"]])) && !all(stub_df[["row_id"]] == "")) {
+  if (!all(is.na(stub_df[["row_id"]])) && !all(!nzchar(stub_df[["row_id"]]))) {
     stub_components <- c(stub_components, "row_id")
   }