--- a/R/dt_stub_df.R
+++ b/R/dt_stub_df.R
@@ -44,7 +44,7 @@ dt_stub_df_init <- function(
 
   # Create the `stub_df` table
   stub_df <-
-    dplyr::tibble(
+    vctrs::data_frame(
       rownum_i = seq_len(nrow(data_tbl)),
       row_id = rep(NA_character_, nrow(data_tbl)),
       group_id = rep(NA_character_, nrow(data_tbl)),