--- a/R/dt_summary.R
+++ b/R/dt_summary.R
@@ -450,8 +450,7 @@ dt_summary_build <- function(data, context) {
       group_summary_display_df <-
         dplyr::filter(summary_dfs_display, .data[[group_id_col_private]] == .env$group)
 
-      group_summary_display_df <-
-        dplyr::mutate(group_summary_display_df, `::side::` = side)
+      group_summary_display_df$`::side::` <- side
 
       summary_df_data_list <-
         c(