--- a/R/gt_group.R
+++ b/R/gt_group.R
@@ -787,7 +787,7 @@ grp_options <- function(
         SIMPLIFY = FALSE
       )
     )
-  new_df <- dplyr::select(new_df, -type)
+  new_df$type <- NULL
 
   # This rearranges the rows in the `opts_df` table, but this
   # shouldn't be a problem