--- a/R/helpers.R
+++ b/R/helpers.R
@@ -3226,7 +3226,7 @@ google_font <- function(name) {
 
   import_stmt <-
     paste_between(
-      gsub(" ", "+", name),
+      gsub(" ", "+", name, fixed = TRUE),
       c(
         "@import url('https://fonts.googleapis.com/css2?family=",
         ":ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap');"