--- a/R/tab_create_modify.R
+++ b/R/tab_create_modify.R
@@ -1044,7 +1044,7 @@ resolve_spanned_column_names <- function(
 #'     fn = function(x) gsub("pct", "%", x)
 #'   ) |>
 #'   text_transform(
-#'     fn = function(x) gsub("\\.", " - ", x),
+#'     fn = function(x) gsub(".", " - ", x, fixed = TRUE),
 #'     locations = cells_column_spanners()
 #'   ) |>
 #'   tab_style(
@@ -3615,8 +3615,8 @@ tab_style <- function(
         # to sentence case
         # cell_grand_summary_row -> grand summary row
         readable_table_part <- attr(loc, "class")[[1]]
-        readable_table_part <- sub("cells_", "", readable_table_part)
-        readable_table_part <- gsub("_", " ", readable_table_part)
+        readable_table_part <- sub("cells_", "", readable_table_part, fixed = TRUE)
+        readable_table_part <- gsub("_", " ", readable_table_part, fixed = TRUE)
 
         cli::cli_abort(
           "Failed to style the {readable_table_part} of the table.",