… R/utils_general_str_formatting.R
… R/utils_plots.R
… R/utils_render_common.R
… R/utils_render_grid.R
… R/utils_render_html.R
… R/utils_render_latex.R
… R/utils_render_rtf.R
… R/utils_render_xml.R
… R/utils_units.R
… R/z_utils_render_footnotes.R
… data-raw/X05-google_fonts.R
… man/tab_spanner_delim.Rd
… tests/testthat/test-as_latex.R
… tests/testthat/test-cols_width.R
… tests/testthat/test-fmt_percent.R
… tests/testthat/test-gt_object.R
… tests/testthat/test-gtsave.R