@rich-iannone 2024-06-21
All of this is so great. I somewhat regret using so much dplyr in the codebase (though I guess it was faster for development) but going down the road of slowly replacing this with base R will make everything so much faster. Thanks for your dedication here!